refactor: split oversized test files to comply with 300-line maximum
Split 5 test files that exceeded the 300-line limit into smaller,
focused modules organized by concern. Created tests/conftest.py with
shared helpers to reduce duplication across split files.

Resolves #42, #65

- test_cli.py (409) → test_cli_doctor.py (92), test_cli_convert.py (148), test_cli_batch.py (143)
- test_docling.py (385) → test_docling_extract.py (256), test_docling_images.py (109)
- test_native_pptx.py (419) → test_native_pptx_extract.py (288), test_native_pptx_tables.py (81)
- test_native_pdf.py (372) → test_native_pdf_extract.py (236), test_native_pdf_images.py (132)
- test_media.py (308) → test_media_process.py (203), test_media_copy.py (100)
