+++ tests/test_native_pdf_extract.py
@@ -1,16 +1,9 @@
-"""Tests for PDF extraction (mock pdfplumber)."""
+"""Tests for PDF text and metadata extraction."""
 
 from pathlib import Path
 from unittest.mock import MagicMock, patch
 
-from hypothesis import given
-from hypothesis import strategies as st
-
-from obsidian_import.backends.native_pdf import (
-    _extract_page_images,
-    _pdf_image_extension,
-    extract,
-)
+from obsidian_import.backends.native_pdf import extract
 from obsidian_import.config import MediaConfig
 from obsidian_import.extraction_result import MediaFile
 
@@ -109,7 +102,6 @@ class TestNativePdfExtract:
         assert "Author Name" in result.markdown
 
     def test_no_images_when_disabled(self, tmp_path):
-        """When extract_images=False, no image extraction is attempted."""
         pdf_path = tmp_path / "test.pdf"
         pdf_path.write_bytes(b"%PDF-1.4 fake")
 
@@ -137,7 +129,6 @@ class TestNativePdfExtract:
         assert result.media_files == ()
 
     def test_wikilinks_use_document_stem_prefix(self, tmp_path):
-        """Wikilinks reference <stem>/<filename>, not media/<filename>."""
         pdf_path = tmp_path / "report.pdf"
         pdf_path.write_bytes(b"%PDF-1.4 fake")
 
@@ -185,7 +176,6 @@ class TestNativePdfExtract:
         assert "![[media/" not in result.markdown
 
     def test_extracts_form_fields(self, tmp_path):
-        """PDF with form fields emits ## Form Fields section."""
         pdf_path = tmp_path / "form.pdf"
         pdf_path.write_bytes(b"%PDF-1.4 fake")
 
@@ -218,7 +208,6 @@ class TestNativePdfExtract:
         assert "**Checkbox1** (/Btn): /Yes" in result.markdown
 
     def test_empty_table_guard(self, tmp_path):
-        """Tables that are empty or have empty first row are skipped."""
         pdf_path = tmp_path / "empty_table.pdf"
         pdf_path.write_bytes(b"%PDF-1.4 fake")
 
@@ -245,128 +234,3 @@ class TestNativePdfExtract:
 
         assert "Header" in result.markdown
         assert "val" in result.markdown
-
-
-class TestExtractPageImages:
-    def test_no_xobjects_returns_empty(self):
-        """Page with resources but no /XObject returns empty list."""
-        mock_resources = MagicMock()
-        mock_resources.get.side_effect = lambda k: None if k == "/XObject" else None
-
-        mock_page = MagicMock()
-        mock_page.get.side_effect = lambda k: mock_resources if k == "/Resources" else None
-
-        mock_reader = MagicMock()
-        mock_reader.pages = [mock_page]
-
-        media_config = MediaConfig(extract_images=True, image_format="png", image_max_dimension=0)
-        result = _extract_page_images(mock_reader, 0, Path("/fake.pdf"), media_config)
-        assert result == []
-
-    def test_non_image_xobject_skipped(self):
-        """XObjects with subtype != /Image are skipped."""
-        mock_xobj = MagicMock()
-        mock_xobj.get.return_value = "/Form"
-
-        mock_xobj_dict = {"form0": MagicMock()}
-        mock_xobj_dict["form0"].get_object.return_value = mock_xobj
-
-        mock_xobjects = MagicMock()
-        mock_xobjects.get_object.return_value = mock_xobj_dict
-
-        mock_resources = MagicMock()
-        mock_resources.get.side_effect = lambda k: mock_xobjects if k == "/XObject" else None
-
-        mock_page = MagicMock()
-        mock_page.get.side_effect = lambda k: mock_resources if k == "/Resources" else None
-
-        mock_reader = MagicMock()
-        mock_reader.pages = [mock_page]
-
-        media_config = MediaConfig(extract_images=True, image_format="png", image_max_dimension=0)
-        result = _extract_page_images(mock_reader, 0, Path("/fake.pdf"), media_config)
-        assert result == []
-
-    def test_image_extraction_failure_logs_warning(self):
-        """When get_data() raises, a warning is logged and processing continues."""
-        mock_xobj = MagicMock()
-        mock_xobj.get.return_value = "/Image"
-        mock_xobj.get_data.side_effect = ValueError("corrupt image")
-
-        mock_xobj_dict = {"img0": MagicMock()}
-        mock_xobj_dict["img0"].get_object.return_value = mock_xobj
-
-        mock_xobjects = MagicMock()
-        mock_xobjects.get_object.return_value = mock_xobj_dict
-
-        mock_resources = MagicMock()
-        mock_resources.get.side_effect = lambda k: mock_xobjects if k == "/XObject" else None
-
-        mock_page = MagicMock()
-        mock_page.get.side_effect = lambda k: mock_resources if k == "/Resources" else None
-
-        mock_reader = MagicMock()
-        mock_reader.pages = [mock_page]
-
-        media_config = MediaConfig(extract_images=True, image_format="png", image_max_dimension=0)
-
-        with patch("obsidian_import.backends.native_pdf.log") as mock_log:
-            result = _extract_page_images(mock_reader, 0, Path("/fake.pdf"), media_config)
-
-        assert result == []
-        mock_log.warning.assert_called_once()
-
-    def test_no_resources_returns_empty(self):
-        """Page with no /Resources returns empty list."""
-        mock_page = MagicMock()
-        mock_page.get.return_value = None
-
-        mock_reader = MagicMock()
-        mock_reader.pages = [mock_page]
-
-        media_config = MediaConfig(extract_images=True, image_format="png", image_max_dimension=0)
-        result = _extract_page_images(mock_reader, 0, Path("/fake.pdf"), media_config)
-        assert result == []
-
-
-class TestPdfImageExtension:
-    def test_none_filter_returns_png(self):
-        xobj = MagicMock()
-        xobj.get.return_value = None
-        assert _pdf_image_extension(xobj) == ".png"
-
-    def test_dctdecode_returns_jpeg(self):
-        xobj = MagicMock()
-        xobj.get.return_value = "/DCTDecode"
-        assert _pdf_image_extension(xobj) == ".jpeg"
-
-    def test_jpxdecode_returns_jp2(self):
-        xobj = MagicMock()
-        xobj.get.return_value = "/JPXDecode"
-        assert _pdf_image_extension(xobj) == ".jp2"
-
-    def test_ccittfaxdecode_returns_tiff(self):
-        xobj = MagicMock()
-        xobj.get.return_value = "/CCITTFaxDecode"
-        assert _pdf_image_extension(xobj) == ".tiff"
-
-    def test_unknown_filter_returns_png(self):
-        xobj = MagicMock()
-        xobj.get.return_value = "/FlateDecode"
-        assert _pdf_image_extension(xobj) == ".png"
-
-    @given(st.sampled_from(["/DCTDecode", "/JPXDecode", "/CCITTFaxDecode", "/FlateDecode", None]))
-    def test_extension_always_starts_with_dot(self, filter_val):
-        """Property: returned extension always starts with a dot."""
-        xobj = MagicMock()
-        xobj.get.return_value = filter_val
-        ext = _pdf_image_extension(xobj)
-        assert ext.startswith(".")
-
-    @given(st.sampled_from(["/DCTDecode", "/JPXDecode", "/CCITTFaxDecode", "/FlateDecode", None]))
-    def test_extension_is_known_format(self, filter_val):
-        """Property: returned extension is one of the known formats."""
-        xobj = MagicMock()
-        xobj.get.return_value = filter_val
-        ext = _pdf_image_extension(xobj)
-        assert ext in {".png", ".jpeg", ".jp2", ".tiff"}
